Follow-up to `enhance login prompt`
