+++ cmd/nerdctl/login_windows.go
@@ -17,10 +17,7 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
-	"os"
-	"strings"
 	"syscall"
 
 	"golang.org/x/term"
@@ -39,22 +36,4 @@ func readPassword() (string, error) {
 	}
 
 	return string(bytePassword), nil
-}
-
-func readUsername() (string, error) {
-	var fd *os.File
-	if term.IsTerminal(int(syscall.Stdin)) {
-		fd = os.Stdin
-	} else {
-		return "", fmt.Errorf("error allocating terminal")
-	}
-
-	reader := bufio.NewReader(fd)
-	username, err := reader.ReadString('\n')
-	if err != nil {
-		return "", fmt.Errorf("error reading username: %w", err)
-	}
-	username = strings.TrimSpace(username)
-
-	return username, nil
 }
